% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\name{apps.create}
\alias{apps.create}
\title{Creates an App Engine application for a Google Cloud Platform project. This requires a project that excludes an App Engine application. For details about creating a project without an application, see the Google Cloud Resource Manager create project topic (https://cloud.google.com/resource-manager/docs/creating-project).}
\usage{
apps.create(Application)
}
\arguments{
\item{Application}{The \link{Application} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/appengine/docs/admin-api/}{Google Documentation}

Other Application functions: \code{\link{Application}},
  \code{\link{apps.patch}}
}
