% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ContainerInfo}
\alias{ContainerInfo}
\title{ContainerInfo Object}
\usage{
ContainerInfo(image = NULL)
}
\arguments{
\item{image}{URI to the hosted container image in a Docker repository}
}
\value{
ContainerInfo object
}
\description{
ContainerInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Docker image that is used to start a VM container for the version you deploy.
}
