% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Deployment}
\alias{Deployment}
\title{Deployment Object}
\usage{
Deployment(Deployment.files = NULL, files = NULL, container = NULL,
  sourceReferences = NULL)
}
\arguments{
\item{Deployment.files}{The \link{Deployment.files} object or list of objects}

\item{files}{Manifest of the files stored in Google Cloud Storage that are included as part of this version}

\item{container}{A Docker image that App Engine uses the run the version}

\item{sourceReferences}{Origin of the source code for this deployment}
}
\value{
Deployment object
}
\description{
Deployment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Code and application artifacts used to deploy a version to App Engine.
}
\seealso{
Other Deployment functions: \code{\link{Deployment.files}}
}

