% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{DiskUtilization}
\alias{DiskUtilization}
\title{DiskUtilization Object}
\usage{
DiskUtilization(targetWriteBytesPerSec = NULL, targetWriteOpsPerSec = NULL,
  targetReadBytesPerSec = NULL, targetReadOpsPerSec = NULL)
}
\arguments{
\item{targetWriteBytesPerSec}{Target bytes written per second}

\item{targetWriteOpsPerSec}{Target ops written per second}

\item{targetReadBytesPerSec}{Target bytes read per second}

\item{targetReadOpsPerSec}{Target ops read per second}
}
\value{
DiskUtilization object
}
\description{
DiskUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by disk usage. Only applicable for VM runtimes.
}

