% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(name = NULL, id = NULL, appEngineRelease = NULL,
  availability = NULL, vmName = NULL, vmZoneName = NULL, vmId = NULL,
  startTimestamp = NULL, requests = NULL, errors = NULL, qps = NULL,
  averageLatency = NULL, memoryUsage = NULL, vmStatus = NULL,
  vmUnlocked = NULL)
}
\arguments{
\item{name}{Full path to the Instance resource in the API}

\item{id}{Relative name of the instance within the version}

\item{appEngineRelease}{App Engine release this instance is running on}

\item{availability}{Availability of the instance}

\item{vmName}{Name of the virtual machine where this instance lives}

\item{vmZoneName}{Zone where the virtual machine is located}

\item{vmId}{Virtual machine ID of this instance}

\item{startTimestamp}{Time that this instance was started}

\item{requests}{Number of requests since this instance was started}

\item{errors}{Number of errors since this instance was started}

\item{qps}{Average queries per second (QPS) over the last minute}

\item{averageLatency}{Average latency (ms) over the last minute}

\item{memoryUsage}{Total memory in use (bytes)}

\item{vmStatus}{Status of the virtual machine where this instance lives}

\item{vmUnlocked}{Whether this instance is in debug mode}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance resource is the computing unit that App Engine uses to automatically scale an application.
}

