% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ListOperationsResponse}
\alias{ListOperationsResponse}
\title{Google App Engine Admin API Objects 
Provisions and manages App Engine applications.}
\usage{
ListOperationsResponse(operations = NULL, nextPageToken = NULL)
}
\arguments{
\item{operations}{A list of operations that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
ListOperationsResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:55:56
filename: /Users/mark/dev/R/autoGoogleAPI/googleappenginev1beta5.auto/R/appengine_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListOperationsResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
