% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ListServicesResponse}
\alias{ListServicesResponse}
\title{ListServicesResponse Object}
\usage{
ListServicesResponse(services = NULL, nextPageToken = NULL)
}
\arguments{
\item{services}{The services belonging to the requested application}

\item{nextPageToken}{Continuation token for fetching the next page of results}
}
\value{
ListServicesResponse object
}
\description{
ListServicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for Services.ListServices.
}
