% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{LocationMetadata}
\alias{LocationMetadata}
\title{LocationMetadata Object}
\usage{
LocationMetadata(standardEnvironmentAvailable = NULL,
  flexibleEnvironmentAvailable = NULL)
}
\arguments{
\item{standardEnvironmentAvailable}{App Engine Standard Environment is available in the given location}

\item{flexibleEnvironmentAvailable}{App Engine Flexible Environment is available in the given location}
}
\value{
LocationMetadata object
}
\description{
LocationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for the given google.cloud.location.Location.
}

