% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{NetworkUtilization}
\alias{NetworkUtilization}
\title{NetworkUtilization Object}
\usage{
NetworkUtilization(targetSentBytesPerSec = NULL,
  targetSentPacketsPerSec = NULL, targetReceivedBytesPerSec = NULL,
  targetReceivedPacketsPerSec = NULL)
}
\arguments{
\item{targetSentBytesPerSec}{Target bytes sent per second}

\item{targetSentPacketsPerSec}{Target packets sent per second}

\item{targetReceivedBytesPerSec}{Target bytes received per second}

\item{targetReceivedPacketsPerSec}{Target packets received per second}
}
\value{
NetworkUtilization object
}
\description{
NetworkUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by network usage. Only applicable for VM runtimes.
}

