% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Resources}
\alias{Resources}
\title{Resources Object}
\usage{
Resources(cpu = NULL, diskGb = NULL, memoryGb = NULL, volumes = NULL)
}
\arguments{
\item{cpu}{Number of CPU cores needed}

\item{diskGb}{Disk size (GB) needed}

\item{memoryGb}{Memory (GB) needed}

\item{volumes}{Volumes mounted within the app container}
}
\value{
Resources object
}
\description{
Resources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Machine resources for a version.
}
