% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{SourceReference}
\alias{SourceReference}
\title{SourceReference Object}
\usage{
SourceReference(repository = NULL, revisionId = NULL)
}
\arguments{
\item{repository}{URI string identifying the repository}

\item{revisionId}{The canonical, persistent identifier of the deployed revision}
}
\value{
SourceReference object
}
\description{
SourceReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Reference to a particular snapshot of the source tree used to build and deploy the application.
}
