% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{StaticFilesHandler}
\alias{StaticFilesHandler}
\title{StaticFilesHandler Object}
\usage{
StaticFilesHandler(StaticFilesHandler.httpHeaders = NULL, path = NULL,
  uploadPathRegex = NULL, httpHeaders = NULL, mimeType = NULL,
  expiration = NULL, requireMatchingFile = NULL,
  applicationReadable = NULL)
}
\arguments{
\item{StaticFilesHandler.httpHeaders}{The \link{StaticFilesHandler.httpHeaders} object or list of objects}

\item{path}{Path to the static files matched by the URL pattern, from the application root directory}

\item{uploadPathRegex}{Regular expression that matches the file paths for all files that should be referenced by this handler}

\item{httpHeaders}{HTTP headers to use for all responses from these URLs}

\item{mimeType}{MIME type used to serve all files served by this handler}

\item{expiration}{Time a static file served by this handler should be cached}

\item{requireMatchingFile}{Whether this handler should match the request if the file referenced by the handler does not exist}

\item{applicationReadable}{Whether files should also be uploaded as code data}
}
\value{
StaticFilesHandler object
}
\description{
StaticFilesHandler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Files served directly to the user for a given URL, such as images, CSS stylesheets, or JavaScript source files. Static file handlers describe which files in the application directory are static files, and which URLs serve them.
}
\seealso{
Other StaticFilesHandler functions: \code{\link{StaticFilesHandler.httpHeaders}}
}

