% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{TrafficSplit.allocations}
\alias{TrafficSplit.allocations}
\title{TrafficSplit.allocations Object}
\usage{
TrafficSplit.allocations()
}
\value{
TrafficSplit.allocations object
}
\description{
TrafficSplit.allocations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mapping from version IDs within the service to fractional (0.000, 1] allocations of traffic for that version. Each version can be specified only once, but some versions in the service may not have any traffic allocation. Services that have traffic allocated cannot be deleted until either the service is deleted or their traffic allocation is removed. Allocations must sum to 1. Up to two decimal place precision is supported for IP-based splits and up to three decimal places is supported for cookie-based splits.
}
\seealso{
Other TrafficSplit functions: \code{\link{TrafficSplit}}
}
