% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{UrlDispatchRule}
\alias{UrlDispatchRule}
\title{UrlDispatchRule Object}
\usage{
UrlDispatchRule(domain = NULL, path = NULL, service = NULL)
}
\arguments{
\item{domain}{Domain name to match against}

\item{path}{Pathname within the host}

\item{service}{Resource id of a service in this application that should serve the matched request}
}
\value{
UrlDispatchRule object
}
\description{
UrlDispatchRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rules to match an HTTP request and dispatch that request to a service.
}
