% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{UrlMap}
\alias{UrlMap}
\title{UrlMap Object}
\usage{
UrlMap(urlRegex = NULL, staticFiles = NULL, script = NULL,
  apiEndpoint = NULL, securityLevel = NULL, login = NULL,
  authFailAction = NULL, redirectHttpResponseCode = NULL)
}
\arguments{
\item{urlRegex}{A URL prefix}

\item{staticFiles}{Returns the contents of a file, such as an image, as the response}

\item{script}{Executes a script to handle the request that matches this URL pattern}

\item{apiEndpoint}{Uses API Endpoints to handle requests}

\item{securityLevel}{Security (HTTPS) enforcement for this URL}

\item{login}{Level of login required to access this resource}

\item{authFailAction}{Action to take when users access resources that require authentication}

\item{redirectHttpResponseCode}{`30x` code to use when performing redirects for the `secure` field}
}
\value{
UrlMap object
}
\description{
UrlMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
URL pattern and description of how the URL should be handled. App Engine can handle URLs by executing application code, or by serving static files uploaded with the version, such as images, CSS, or JavaScript.
}

