% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Version.betaSettings}
\alias{Version.betaSettings}
\title{Version.betaSettings Object}
\usage{
Version.betaSettings()
}
\value{
Version.betaSettings object
}
\description{
Version.betaSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata settings that are supplied to this version to enable beta runtime features.
}
\seealso{
Other Version functions: \code{\link{Version.envVariables}},
  \code{\link{Version}}
}
