% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Version}
\alias{Version}
\title{Version Object}
\usage{
Version(Version.betaSettings = NULL, Version.envVariables = NULL,
  name = NULL, id = NULL, automaticScaling = NULL, basicScaling = NULL,
  manualScaling = NULL, inboundServices = NULL, instanceClass = NULL,
  network = NULL, resources = NULL, runtime = NULL, threadsafe = NULL,
  vm = NULL, betaSettings = NULL, env = NULL, servingStatus = NULL,
  deployer = NULL, creationTime = NULL, diskUsageBytes = NULL,
  handlers = NULL, errorHandlers = NULL, libraries = NULL,
  apiConfig = NULL, envVariables = NULL, defaultExpiration = NULL,
  healthCheck = NULL, nobuildFilesRegex = NULL, deployment = NULL)
}
\arguments{
\item{Version.betaSettings}{The \link{Version.betaSettings} object or list of objects}

\item{Version.envVariables}{The \link{Version.envVariables} object or list of objects}

\item{name}{Full path to the Version resource in the API}

\item{id}{Relative name of the version within the module}

\item{automaticScaling}{Automatic scaling is based on request rate, response latencies, and other application metrics}

\item{basicScaling}{A service with basic scaling will create an instance when the application receives a request}

\item{manualScaling}{A service with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time}

\item{inboundServices}{Before an application can receive email or XMPP messages, the application must be configured to enable the service}

\item{instanceClass}{Instance class that is used to run this version}

\item{network}{Extra network settings}

\item{resources}{Machine resources for this version}

\item{runtime}{Desired runtime}

\item{threadsafe}{Whether multiple requests can be dispatched to this version at once}

\item{vm}{Whether to deploy this version in a container on a virtual machine}

\item{betaSettings}{Metadata settings that are supplied to this version to enable beta runtime features}

\item{env}{App Engine execution environment to use for this version}

\item{servingStatus}{Current serving status of this version}

\item{deployer}{Email address of the user who created this version}

\item{creationTime}{Time that this version was created}

\item{diskUsageBytes}{Total size of version files hosted on App Engine disk in bytes}

\item{handlers}{An ordered list of URL-matching patterns that should be applied to incoming requests}

\item{errorHandlers}{Custom static error pages}

\item{libraries}{Configuration for third-party Python runtime libraries required by the application}

\item{apiConfig}{Serving configuration for [Google Cloud Endpoints](https://cloud}

\item{envVariables}{Environment variables made available to the application}

\item{defaultExpiration}{Duration that static files should be cached by web proxies and browsers}

\item{healthCheck}{Configures health checking for VM instances}

\item{nobuildFilesRegex}{Files that match this pattern will not be built into this version}

\item{deployment}{Code and application artifacts that make up this version}
}
\value{
Version object
}
\description{
Version Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Version resource is a specific set of source code and configuration files that are deployed into a service.
}
\seealso{
Other Version functions: \code{\link{Version.betaSettings}},
  \code{\link{Version.envVariables}}
}

